pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-tutor"
version = "0.1.0"
description = "CLI scaffold for an AI tutoring assistant."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "AI Tutor Team" }
]
dependencies = [
  "pyyaml>=6.0",
  "python-dotenv>=1.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "ruff>=0.5.0"
]

[project.scripts]
ai-tutor = "ai_tutor.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 88
